+++ Source/Diagnostics/ReducedDiags/MultiReducedDiags.cpp
@@ -34,7 +34,7 @@ MultiReducedDiags::MultiReducedDiags ()
     m_multi_rd.resize(m_rd_names.size());
 
     // loop over all reduced diags
-    for (int i_rd = 0; i_rd < m_rd_names.size(); ++i_rd)
+    for (int i_rd = 0; i_rd < static_cast<int>(m_rd_names.size()); ++i_rd)
     {
 
         ParmParse pp_rd(m_rd_names[i_rd]);
@@ -83,7 +83,7 @@ MultiReducedDiags::MultiReducedDiags ()
 void MultiReducedDiags::ComputeDiags (int step)
 {
     // loop over all reduced diags
-    for (int i_rd = 0; i_rd < m_rd_names.size(); ++i_rd)
+    for (int i_rd = 0; i_rd < static_cast<int>(m_rd_names.size()); ++i_rd)
     {
         m_multi_rd[i_rd] -> ComputeDiags(step);
     }
@@ -99,7 +99,7 @@ void MultiReducedDiags::WriteToFile (int step)
     if ( !ParallelDescriptor::IOProcessor() ) { return; }
 
     // loop over all reduced diags
-    for (int i_rd = 0; i_rd < m_rd_names.size(); ++i_rd)
+    for (int i_rd = 0; i_rd < static_cast<int>(m_rd_names.size()); ++i_rd)
     {
 
         // Judge if the diags should be done
